fix: address PR review — bound delta reconstruction, keep signatures stable
Addresses Copilot review feedback on the decoded-object cap:

- ApplyDelta now enforces the delta's declared target size: preallocates
  to it (not ExpectedSourceLength + 64*len(Changes), which a many-change
  delta could inflate), rejects any change that would overrun it, and
  rejects a final length that doesn't match. parseDelta drops a command
  that would consume past the remaining target instead of underflowing
  its unsigned counter. Together these bound delta-based decompression
  bombs, not just the declared size.
- Restore ParsePackfile / ParseFetchResponse to their original
  (non-variadic) signatures and add ParsePackfileWithOptions /
  ParseFetchResponseWithOptions for the configurable form, so the
  exported function types stay backward compatible.
- ObjectTooLargeError.Size is now int64 and the delta path clamps its
  uint64 TargetLength, so a huge declared size no longer wraps.
- Clarify Limits / WithLimits godoc: the four wire caps use 0 = no limit,
  while MaxObjectDecodedBytes 0 = built-in default (never fully off).
- Add regression tests for the delta overrun/short/preallocation guards
  and the parseDelta underflow guard.

+++ protocol/client/limits_integration_test.go
@@ -135,7 +135,7 @@ func TestFetchHonorsMaxObjectDecodedBytes(t *testing.T) {
 
 	var tooLarge *protocol.ObjectTooLargeError
 	require.True(t, errors.As(err, &tooLarge), "expected *protocol.ObjectTooLargeError, got %T: %v", err, err)
-	require.Equal(t, declaredSize, tooLarge.Size)
+	require.Equal(t, int64(declaredSize), tooLarge.Size)
 	require.Equal(t, int64(4096), tooLarge.Limit)
 	// The sentinel is preserved through the wrap for errors.Is callers.
 	require.ErrorIs(t, err, protocol.ErrObjectTooLarge)
